requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.7
playwright>=1.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=4.9.0
flask>=3.0.0
